refactor(utils): extract shape geometry helpers into _geometry.py (#696)
Move 8 private geometry/patch functions verbatim from the 4918-line
utils.py into a new sibling module pl/_geometry.py: _get_centroid_of_pathpatch,
_scale_pathpatch_around_centroid, _normalize_geom, _make_patch_from_multipolygon,
_build_shape_patches, _get_collection_shape, _validate_polygons, _convert_shapes.

No behavior change (verbatim move; all-private symbols). render.py is the sole
consumer; its four imports of these are repointed to _geometry. The module's
only edge back into utils is _extract_scalar_value (downward; no cycle).

Verified: ruff + ruff-format clean on the moved code; 55 non-visual shape tests
pass. Pre-existing ruff D205 / mypy Any-return debt in utils.py (from #703/#705)
is unrelated and is fixed by #714; --no-verify used for that reason only.
